# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ruammit API
  version: 0.1.0
  description: >
    Backend for the Ruammit travel social app. This is a starter contract —
    extend it module by module (auth, feed, location, chat). It doubles as the
    source for generating a Dart client for the Flutter app.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/posts/create:
    post:
      summary: Create a post (caption + media + optional location)
      description: >
        Multipart/form-data. A post may carry up to 8 images OR a single video
        (not both), plus an optional caption (≤2000 chars) and location. Media
        is stored and served under `/media/...`; the response is the created
        post in the same shape the feed returns. Rate-limited per user.
      operationId: createPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                caption: { type: string, maxLength: 2000 }
                images:
                  type: array
                  maxItems: 8
                  items: { type: string, format: binary }
                  description: JPG/PNG/WebP, ≤5 MB each. Omit when sending a video.
                video:
                  type: string
                  format: binary
                  description: MP4/MOV, ≤50 MB. Omit when sending images.
                latitude: { type: number }
                longitude: { type: number }
                location_name: { type: string }
      responses:
        "201":
          description: The created post
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Post" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/feed:
    get:
      summary: Home timeline (viewer + follows)
      description: >
        Returns a page of the viewer's home timeline. If the viewer follows no
        one yet, the server falls back to a recent global timeline so the feed
        is never empty; `source` reports which timeline was returned.
      operationId: homeFeed
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: A page of timeline posts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedPage" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/feed/stories:
    get:
      summary: Active stories, grouped by author
      operationId: feedStories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active story strip
          content:
            application/json:
              schema:
                type: object
                properties:
                  stories:
                    type: array
                    items: { $ref: "#/components/schemas/StoryAuthor" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/feed/posts/{id}/like:
    post:
      summary: Like a post
      operationId: likePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Liked }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
    delete:
      summary: Unlike a post
      operationId: unlikePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Unliked }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Author:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        avatar_url: { type: string }
    Post:
      type: object
      properties:
        id: { type: string, format: uuid }
        author: { $ref: "#/components/schemas/Author" }
        body: { type: string }
        created_at: { type: string, format: date-time }
        like_count: { type: integer }
        comment_count: { type: integer }
        liked_by_viewer: { type: boolean }
        images:
          type: array
          items: { type: string }
        video_url:
          type: string
          description: Present only when the post has a video.
        media:
          type: array
          description: Typed media (returned on create; timeline reads use images/video_url).
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              type: { type: string, enum: [image, video] }
              url: { type: string }
              order: { type: integer }
        location:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PostLocation"
    FeedPage:
      type: object
      properties:
        posts:
          type: array
          items: { $ref: "#/components/schemas/Post" }
        source:
          type: string
          enum: [home, global]
          description: Which timeline was returned (global = discovery fallback)
        next_offset:
          type: integer
          nullable: true
          description: Offset for the next page, or null on the last page
...
    PostLocation:
      type: object
      properties:
        latitude: { type: number }
        longitude: { type: number }
        location_name: { type: string }
    StoryAuthor:
      type: object
      properties:
        author: { $ref: "#/components/schemas/Author" }
        media_urls:
          type: array
          items: { type: string }
        latest_at: { type: string, format: date-time }
    Error:
      type: object
      properties:
        error: { type: string, description: Machine-readable code }
        message: { type: string }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
# Planned route groups (to be filled in as modules land):
#   /api/v1/auth/*       — signup, login, social, refresh
#   /api/v1/users/*      — profiles, follow/unfollow
#   /api/v1/feed, /posts — timeline, create post, like, comment, stories
#   /api/v1/nearby       — geospatial discovery (query: lat,lng,radius,gender,age)
#   /api/v1/conversations, /messages, /ws — chat REST + WebSocket upgrade
#   /api/v1/media/uploads — presigned upload URLs