# internal/platform/storage/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package dbgen

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Comment struct {
	ID        pgtype.UUID        `json:"id"`
	PostID    pgtype.UUID        `json:"post_id"`
	AuthorID  pgtype.UUID        `json:"author_id"`
	Body      string             `json:"body"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type Conversation struct {
	ID        pgtype.UUID        `json:"id"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type ConversationMember struct {
	ConversationID pgtype.UUID `json:"conversation_id"`
	UserID         pgtype.UUID `json:"user_id"`
}

type Device struct {
	ID        pgtype.UUID        `json:"id"`
	UserID    pgtype.UUID        `json:"user_id"`
	FcmToken  string             `json:"fcm_token"`
	Platform  string             `json:"platform"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type Follow struct {
	FollowerID pgtype.UUID        `json:"follower_id"`
	FolloweeID pgtype.UUID        `json:"followee_id"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type Message struct {
	ID             pgtype.UUID        `json:"id"`
	ConversationID pgtype.UUID        `json:"conversation_id"`
	SenderID       pgtype.UUID        `json:"sender_id"`
	Kind           string             `json:"kind"`
	Body           *string            `json:"body"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
}

type OtpCode struct {
	ID         pgtype.UUID        `json:"id"`
	UserID     pgtype.UUID        `json:"user_id"`
	Purpose    string             `json:"purpose"`
	CodeHash   string             `json:"code_hash"`
	Attempts   int32              `json:"attempts"`
	ExpiresAt  pgtype.Timestamptz `json:"expires_at"`
	ConsumedAt pgtype.Timestamptz `json:"consumed_at"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type Post struct {
	ID          pgtype.UUID        `json:"id"`
	AuthorID    pgtype.UUID        `json:"author_id"`
	Body        string             `json:"body"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	MediaCount  int16              `json:"media_count"`
	HasLocation bool               `json:"has_location"`
}

type PostImage struct {
	ID       pgtype.UUID `json:"id"`
	PostID   pgtype.UUID `json:"post_id"`
	Url      string      `json:"url"`
	Position int16       `json:"position"`
}

type PostLike struct {
	PostID    pgtype.UUID        `json:"post_id"`
	UserID    pgtype.UUID        `json:"user_id"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type PostLocation struct {
	ID           pgtype.UUID        `json:"id"`
	PostID       pgtype.UUID        `json:"post_id"`
	Latitude     float64            `json:"latitude"`
	Longitude    float64            `json:"longitude"`
	LocationName *string            `json:"location_name"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
}

type PostMedium struct {
	ID         pgtype.UUID        `json:"id"`
	PostID     pgtype.UUID        `json:"post_id"`
	MediaType  string             `json:"media_type"`
	MediaUrl   string             `json:"media_url"`
	MediaOrder int16              `json:"media_order"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type RefreshToken struct {
	ID        pgtype.UUID        `json:"id"`
	UserID    pgtype.UUID        `json:"user_id"`
	TokenHash string             `json:"token_hash"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	RevokedAt pgtype.Timestamptz `json:"revoked_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type SocialIdentity struct {
	ID          pgtype.UUID        `json:"id"`
	UserID      pgtype.UUID        `json:"user_id"`
	Provider    string             `json:"provider"`
	ProviderUid string             `json:"provider_uid"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type Story struct {
	ID        pgtype.UUID        `json:"id"`
	AuthorID  pgtype.UUID        `json:"author_id"`
	MediaUrl  string             `json:"media_url"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
}

type User struct {
	ID           pgtype.UUID        `json:"id"`
	Email        *string            `json:"email"`
	PasswordHash *string            `json:"password_hash"`
	DisplayName  *string            `json:"display_name"`
	AvatarUrl    *string            `json:"avatar_url"`
	Bio          *string            `json:"bio"`
	Gender       *string            `json:"gender"`
	BirthDate    pgtype.Date        `json:"birth_date"`
	Location     string             `json:"location"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
	Phone        *string            `json:"phone"`
	Status       string             `json:"status"`
	VerifiedAt   pgtype.Timestamptz `json:"verified_at"`
}